pyproject.toml
[project]
name = "lipvm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "debugpy>=1.8.19",
]

[tool.uv.workspace]
members = [
    "src/lipvm",
]
